+++ apps/dine-pleiepenger/src/server/api-models/k9FormatEttersendelseSchema.ts
@@ -5,5 +5,5 @@ import { Ettersendelsestype } from '../../types/EttersendelseType';
 export const K9FormatEttersendelseSchema = z.object({
     søknadId: z.string(),
     mottattDato: z.preprocess(parseMaybeDateStringToDate, z.date()),
-    type: z.nativeEnum(Ettersendelsestype),
+    type: z.nativeEnum(Ettersendelsestype).optional().nullable(),
 });